添加my_types.h头文件，重构GPIO和LED模块，统一数据类型为uchar，修改goip_Init()函数命名

+++ StaticDigital/include/led.h
@@ -2,15 +2,16 @@
 #define _MYLED_H
 
 #include <8052.h>
+#include "my_types.h"
 
 // 静态点亮数码管函数声明，position: 数码管位置（0-2），number: 数字（0-9）
-void static_led_on(int position, int number);
+void static_led_on(uchar position, uchar number);
 
 // 关闭数码管函数声明，position: 数码管位置（0-2）
-void led_off(int position);
+void led_off(uchar position);
 
 // 动态点亮数码管函数声明，position: 数码管位置（0-2），number: 数字（0-9）
-void dynamic_led_on(int number_one, int number_two, int number_three);
+void dynamic_led_on(uchar number_one, uchar number_two, uchar number_three);
 
 #endif
 // 数码管控制函数声明
